Makefile: Add VC option so we can use this plugin on other platforms
Raspberry PI 2 is no longer "armv6l" (it is "armv7l"). An option is more flexible as a hard coded MACHINE test. So we can use this plugin on other broadcom platforms as well. I have not removed ("$(MACHINE)","armv6l") because it would break backward compatibility with your mupen64plus build script.